Fixed by comments
