+++ hw2/src/main/java/animals/Crocodile.java
@@ -4,13 +4,6 @@ import interfaces.Swim;
 import interfaces.Voice;
 
 public class Crocodile extends Carnivore implements Swim, Voice {
-    public void swim() {
-        System.out.println(this.name + " can swim!");
-    }
-
-    public String voice() {
-        return "Rrrrr!";
-    }
 
     public Crocodile() {
     }
@@ -22,24 +15,12 @@ public class Crocodile extends Carnivore implements Swim, Voice {
         this.color = color;
     }
 
-    public String getName() {
-        return name;
+    public void swim() {
+        System.out.println(this.name + " can swim!");
     }
 
-    public double getWeight() {
-        return weight;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public String typeAnimal() {
-        return "Crocodile";
+    public String voice() {
+        return "Rrrrr!";
     }
 
     public void addSatiety() {
