+++ hw2/src/main/java/animals/Duck.java
@@ -6,6 +6,17 @@ import interfaces.Swim;
 import interfaces.Voice;
 
 public class Duck extends Herbivore implements Fly, Run, Swim, Voice {
+
+    public Duck() {
+    }
+
+    public Duck(String name, double weight, int age, String color) {
+        this.name = name;
+        this.weight = weight;
+        this.age = age;
+        this.color = color;
+    }
+
     public void fly() {
         System.out.println(this.name + " can fly!");
     }
@@ -20,36 +31,6 @@ public class Duck extends Herbivore implements Fly, Run, Swim, Voice {
 
     public String voice() {
         return "Quack!";
-    }
-
-    public Duck() {
-    }
-
-    public Duck(String name, double weight, int age, String color) {
-        this.name = name;
-        this.weight = weight;
-        this.age = age;
-        this.color = color;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public double getWeight() {
-        return weight;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public String typeAnimal() {
-        return "Duck";
     }
 
     public void addSatiety() {
