+++ hw2/src/main/java/animals/Wolf.java
@@ -5,17 +5,6 @@ import interfaces.Swim;
 import interfaces.Voice;
 
 public class Wolf extends Carnivore implements Run, Swim, Voice {
-    public void run() {
-        System.out.println(this.name + " can run!");
-    }
-
-    public void swim() {
-        System.out.println(this.name + " can swim!");
-    }
-
-    public String voice() {
-        return "Woooow!";
-    }
 
     public Wolf() {
     }
@@ -27,24 +16,16 @@ public class Wolf extends Carnivore implements Run, Swim, Voice {
         this.color = color;
     }
 
-    public String getName() {
-        return name;
+    public void run() {
+        System.out.println(this.name + " can run!");
     }
 
-    public double getWeight() {
-        return weight;
+    public void swim() {
+        System.out.println(this.name + " can swim!");
     }
 
-    public int getAge() {
-        return age;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public String typeAnimal() {
-        return "Wolf";
+    public String voice() {
+        return "Woooow!";
     }
 
     public void addSatiety() {
